requirements.txt
docker
gunicorn
honcho
python-dotenv
websocket_client

flask
flask-dance
flask-sockets

flake8
mock
pycodestyle
pylint
stestr
